requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
pyyaml>=6.0
